temporary workaround for compiler regression
svn path=/trunk/mcs/; revision=58385

+++ mcs/class/System.Web/System.Web.Configuration_2.0/CacheSection.cs
@@ -91,7 +91,7 @@ namespace System.Web.Configuration {
 			set { base[percentagePhysicalMemoryUsedLimitProp] = value; }
 		}
 
-		[LongValidator (MinValue = 0, MaxValue = Int64.MaxValue)]
+		[LongValidator (MinValue = (long) 0, MaxValue = Int64.MaxValue)]
 		[ConfigurationProperty ("privateBytesLimit", DefaultValue = "0")]
 		public long PrivateBytesLimit {
 			get { return (long) base [privateBytesLimitProp];}
